pyproject.toml
[project]
name = "primeiro-agente"
version = "0.1.0"
description = "primeiro agente inteligente"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "agno>=1.5.6",
    "arxiv>=2.2.0",
    "googlesearch-python>=1.3.0",
    "ipykernel>=6.29.5",
    "notebook>=7.4.3",
    "ollama>=0.5.1",
    "openai>=1.83.0",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pycountry>=24.6.1",
    "pymupdf>=1.26.0",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
]
[tool.taskipy.tasks]
gitadd = "git add ."
gitcommit = "git commit -m 'update'"
gitpush = "git push origin main"
deploy = "ssh root@192.168. 'cd /caminho/do/repositorio && git pull'"
pushdeploy = "task gitadd && task gitcommit && task gitpush && task deploy"
psh = "task gitadd && task gitcommit && task gitpush "

[dependency-groups]
dev = [
    "notebook>=7.4.3",
    "taskipy>=1.14.1",
]
